Updated vulnerability detail view and added feature "Findings". Updated all list views with genericFilters and fixed small bugs
