+++ occtet-frontend/src/main/java/eu/occtet/bocfrontend/view/softwareComponent/SoftwareComponentDetailView.java
@@ -39,10 +39,9 @@ import eu.occtet.bocfrontend.entity.Project;
 import eu.occtet.bocfrontend.entity.SoftwareComponent;
 import eu.occtet.bocfrontend.entity.Vulnerability;
 import eu.occtet.bocfrontend.service.NatsService;
-import eu.occtet.bocfrontend.view.audit.filestabfragment.FilesTabFragment;
 import eu.occtet.bocfrontend.view.dialog.AddLicenseDialog;
 import eu.occtet.bocfrontend.view.main.MainView;
-import eu.occtet.bocfrontend.view.vulnerabilitydetail.VulnerabilityDetailView;
+import eu.occtet.bocfrontend.view.vulnerability.VulnerabilityDetailView;
 import io.jmix.flowui.DialogWindows;
 import io.jmix.flowui.Notifications;
 import io.jmix.flowui.UiComponents;
