+++ occtet-frontend/src/main/java/eu/occtet/bocfrontend/view/vulnerability/VulnerabilityListView.java
@@ -25,6 +25,7 @@ package eu.occtet.bocfrontend.view.vulnerability;
 import com.vaadin.flow.router.Route;
 import eu.occtet.bocfrontend.entity.Vulnerability;
 import eu.occtet.bocfrontend.view.main.MainView;
+
 import io.jmix.flowui.view.*;
 
 
@@ -34,4 +35,6 @@ import io.jmix.flowui.view.*;
 @LookupComponent("vulnerabilitiesDataGrid")
 @DialogMode(width = "64em")
 public class VulnerabilityListView extends StandardListView<Vulnerability> {
+
+
 }
