--- occtet-frontend/src/main/java/eu/occtet/bocfrontend/view/vulnerabilitydetail/VulnerabilityDetailView.java
@@ -1,72 +1,0 @@
-/*
- *
- *  Copyright (C) 2025 Bitsea GmbH
- *  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- *  SPDX-License-Identifier: Apache-2.0
- *  License-Filename: LICENSE
- * /
- *
- */
-
-package eu.occtet.bocfrontend.view.vulnerabilitydetail;
-
-
-import com.vaadin.flow.component.textfield.TextArea;
-import com.vaadin.flow.router.Route;
-import eu.occtet.bocfrontend.entity.Vulnerability;
-import eu.occtet.bocfrontend.view.main.MainView;
-import io.jmix.flowui.view.*;
-
-import java.util.stream.Collectors;
-
-@Route(value = "vulnerability-detail-view", layout = MainView.class)
-@ViewController(id = "VulnerabilityDetailView")
-@ViewDescriptor(path = "vulnerability-detail-view.xml")
-@EditedEntityContainer("vulnerabilityDc")
-public class VulnerabilityDetailView extends StandardDetailView<Vulnerability> {
-
-    @ViewComponent
-    private TextArea aliasesField;
-
-    @ViewComponent
-    private TextArea fixedPackagesField;
-
-    @ViewComponent
-    private TextArea referencesField;
-
-    @Subscribe
-    public void onBeforeShow(BeforeShowEvent event) {
-        Vulnerability vuln = getEditedEntity();
-
-        aliasesField.setValue(
-                (vuln.getAliases() == null || vuln.getAliases().isEmpty())
-                        ? ""
-                        : String.join("\n", vuln.getAliases())
-        );
-
-        fixedPackagesField.setValue(
-                (vuln.getFixedPackages() == null || vuln.getFixedPackages().isEmpty())
-                        ? ""
-                        : String.join("\n", vuln.getFixedPackages())
-        );
-
-        referencesField.setValue(
-                (vuln.getReferences() == null || vuln.getReferences().isEmpty())
-                        ? ""
-                        : String.join("\n", vuln.getReferences())
-        );
-    }
-
-}
