+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/messages_en.properties
@@ -94,7 +94,7 @@ eu.occtet.bocfrontend/menu.application.title=Administration
 eu.occtet.bocfrontend.entity/CodeLocation=Code location
 eu.occtet.bocfrontend.entity/CodeLocation.filePath=File path
 eu.occtet.bocfrontend.entity/CodeLocation.id=Id
-eu.occtet.bocfrontend.entity/LicenseType=License type
+eu.occtet.bocfrontend.entity/License=License type
 eu.occtet.bocfrontend.entity/LicenseType.detailsUrl=URL
 eu.occtet.bocfrontend.entity/LicenseType.id=Id
 eu.occtet.bocfrontend.entity/LicenseType.isDeprecatedLicenseId=Deprecated
@@ -125,9 +125,8 @@ dateTimeFormat = MM/dd/yyyy HH:mm:ss:SSS
 eu.occtet.bocfrontend/ScannerInitializer.running=Running Scanners\
 
 
+eu.occtet.bocfrontend.view.vulnerability/vulnerabilityDetailView.title=Vulnerability 
 eu.occtet.bocfrontend.view.vulnerability/vulnerabilityListView.title=Vulnerabilities
-
-eu.occtet.bocfrontend.view.vulnerabilitydetail/vulnerabilityDetailView.title=Vulnerability detail view title
 
 eu.occtet.bocfrontend/ScannerInitializer.available=Available scanners
 eu.occtet.bocfrontend/ScannerInitializer.scannerInitializer=Scanner Initializer
@@ -169,9 +168,10 @@ eu.occtet.bocfrontend.view.softwareComponent/softwareComponent.licenses=Licenses
 eu.occtet.bocfrontend.view.softwareComponent/softwareComponent.inventoryItem=Inventory item
 
 eu.occtet.bocfrontend.entity/Project.id=Id
+eu.occtet.bocfrontend.entity/Project.projectName=Name
 eu.occtet.bocfrontend.entity/softwareComponent=Software component
-eu.occtet.bocfrontend.entity/softwareComponent.id=Id
-eu.occtet.bocfrontend.entity/softwareComponent.name=Software Component name
+eu.occtet.bocfrontend.entity/SoftwareComponent.id=Id
+eu.occtet.bocfrontend.entity/SoftwareComponent.name=Name
 eu.occtet.bocfrontend.entity/softwareComponent.project=Project
 eu.occtet.bocfrontend.entity/softwareComponent.basePath=Basepath
 eu.occtet.bocfrontend.entity/Project.name=Name
@@ -193,4 +193,54 @@ eu.occtet.bocfrontend.entity/softwareComponent.parentId=Parent id
 eu.occtet.bocfrontend.entity/ScannerInitializer.status=Status
 eu.occtet.bocfrontend.entity/ScannerInitializer.project=Project
 
+eu.occtet.bocfrontend.entity/Copyright.copyrightText=Name
+eu.occtet.bocfrontend.entity/Copyright.curated=Curated
+eu.occtet.bocfrontend.entity/Copyright.garbage=Garbage
+eu.occtet.bocfrontend.entity/Copyright.id=ID
+eu.occtet.bocfrontend.entity/Copyright.codeLocation=Codelocation
 
+eu.occtet.bocfrontend.entity/SoftwareComponent.detailsUrl=Detail URL
+eu.occtet.bocfrontend.entity/SoftwareComponent.vulnerabilities=Vulnerabilities
+eu.occtet.bocfrontend.entity/SoftwareComponent.purl=PURL
+eu.occtet.bocfrontend.entity/SoftwareComponent.curated=Curated
+eu.occtet.bocfrontend.entity/SoftwareComponent.licenses=Licenses
+eu.occtet.bocfrontend.entity/SoftwareComponent.version=Version
+
+eu.occtet.bocfrontend.entity/License.detailsUrl=URL
+eu.occtet.bocfrontend.entity/License.id=Id
+eu.occtet.bocfrontend.entity/License.priority=Priority
+eu.occtet.bocfrontend.entity/License.licenseType=Type
+eu.occtet.bocfrontend.entity/License.licenseText=License text
+eu.occtet.bocfrontend.entity/License.licenseName=License name
+eu.occtet.bocfrontend.entity/License.isModified=Modified
+eu.occtet.bocfrontend.entity/License.name=Name
+eu.occtet.bocfrontend.entity/License.curated=Curated
+eu.occtet.bocfrontend.entity/License.isSpdx=SPDX
+
+eu.occtet.bocfrontend.entity/InventoryItem.id=ID
+eu.occtet.bocfrontend.entity/InventoryItem.inventoryName=Name
+eu.occtet.bocfrontend.entity/InventoryItem.size=Size
+eu.occtet.bocfrontend.entity/InventoryItem.spdxId=SPDX-ID
+eu.occtet.bocfrontend.entity/InventoryItem.linking=Linking
+eu.occtet.bocfrontend.entity/InventoryItem.priority=Priority
+eu.occtet.bocfrontend.entity/InventoryItem.conspicuous=Conspicuous
+eu.occtet.bocfrontend.entity/InventoryItem.copyrights=Copyrights
+eu.occtet.bocfrontend.entity/InventoryItem.externalNotes=Notes
+eu.occtet.bocfrontend.entity/InventoryItem.parent=InventoryItem-parent
+eu.occtet.bocfrontend.entity/InventoryItem.softwareComponent=SoftwareComponent
+eu.occtet.bocfrontend.entity/InventoryItem.wasCombined=Combined
+eu.occtet.bocfrontend.entity/InventoryItem.curated=Curated
+eu.occtet.bocfrontend.entity/InventoryItem.project=Project
+eu.occtet.bocfrontend.entity/InventoryItem.basePath=Basepath
+eu.occtet.bocfrontend.entity/InventoryItem.createdAt=Created
+
+eu.occtet.bocfrontend.entity/Vulnerability.vulnerabilityId=ID
+eu.occtet.bocfrontend.entity/Vulnerability.summary=Summary
+eu.occtet.bocfrontend.entity/Vulnerability.aliases=Aliases
+eu.occtet.bocfrontend.entity/Vulnerability.weightedSeverity=Serverity
+eu.occtet.bocfrontend.entity/Vulnerability.exploitability=Exploitability
+eu.occtet.bocfrontend.entity/Vulnerability.riskScore=Riskscore
+eu.occtet.bocfrontend.entity/Vulnerability.fixedPackages=Fixed packages
+eu.occtet.bocfrontend.entity/Vulnerability.sourceUrl=Source url
+eu.occtet.bocfrontend.entity/Vulnerability.references=References
+eu.occtet.bocfrontend.entity/Vulnerability.createdAt=Created
