+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/copyright/copyright-list-view.xml
@@ -50,33 +50,17 @@
     </actions>
     <layout>
         <vbox>
-            <accordion id="projectAccordion" width="100%" >
-                <accordionPanel summaryText="Find copyrights from software component">
-                    <hbox>
-                        <comboBox id="projectComboBox" label="Project" width="20em"/>
-                        <comboBox id="softwareComponentComboBox" label="Software component" width="20em"/>
-                    </hbox>
-                </accordionPanel>
-            </accordion>
-            <accordion id="attributesAccordion" width="100%">
-                <accordionPanel summaryText="Filter copyrights by attributes">
-                    <hbox>
-                        <comboBox id="curatedComboBox" label="Curated"/>
-                        <comboBox id="garbageComboBox" label="Garbage"/>
-                    </hbox>
-                </accordionPanel>
-            </accordion>
-            <hbox classNames="buttons-panel" width="100%">
-                <hbox alignSelf="END">
-                    <textField id="searchField" width="15em" placeholder="search term"/>
-                    <button id="searchButton" text="Search"/>
-                </hbox>
-                <button id="refreshButton" action="copyrightsDataGrid.refreshAction" alignSelf="CENTER"/>
+            <hbox width="100%">
+                <genericFilter id="genericFilter" dataLoader="copyrightsDl" themeNames="small"
+                               width="80%" alignSelf="START">
+                    <properties include=".*"/>
+                    <conditions> </conditions>
+                </genericFilter>
                 <simplePagination id="pagination" dataLoader="copyrightsDl" alignSelf="END"/>
             </hbox>
             <dataGrid id="copyrightsDataGrid"
                       width="100%"
-                      minHeight="20em"
+                      minHeight="30em"
                       dataContainer="copyrightsDc"
                       columnReorderingAllowed="true">
                 <columns resizable="true">
