+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/license/license-list-view.xml
@@ -48,33 +48,17 @@
     </actions>
     <layout>
         <vbox>
-            <accordion id="projectAccordion" width="100%" >
-                <accordionPanel summaryText="Find licenses from software component">
-                    <hbox>
-                        <comboBox id="projectComboBox" label="Project" width="20em" />
-                        <comboBox id="softwareComponentComboBox" label="Software component" width="20em"/>
-                    </hbox>
-                </accordionPanel>
-            </accordion>
-            <accordion id="attributesAccordion" width="100%">
-                <accordionPanel summaryText="Filter licenses by attributes">
-                    <hbox>
-                        <comboBox id="priorityComboBox" label="Priority"/>
-                        <comboBox id="curatedComboBox" label="Curated"/>
-                    </hbox>
-                </accordionPanel>
-            </accordion>
-            <hbox classNames="buttons-panel" width="100%">
-                <hbox alignSelf="END">
-                    <textField id="searchField" width="15em" placeholder="search term"/>
-                    <button id="searchButton" text="Search"/>
-                </hbox>
-                <button id="refreshButton" action="licensesDataGrid.refreshAction" alignSelf="CENTER"/>
+            <hbox width="100%">
+                <genericFilter id="genericFilter" dataLoader="licensesDl" themeNames="small"
+                               width="80%" alignSelf="START">
+                    <properties include=".*"/>
+                    <conditions> </conditions>
+                </genericFilter>
                 <simplePagination id="pagination" dataLoader="licensesDl" alignSelf="END"/>
             </hbox>
             <dataGrid id="licensesDataGrid"
                       width="100%"
-                      height="35em"
+                      height="30em"
                       dataContainer="licensesDc"
                       columnReorderingAllowed="true">
                 <actions>
