+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/project/project-list-view.xml
@@ -48,27 +48,26 @@
     </actions>
     <layout>
         <vbox width="100%">
-            <hbox classNames="buttons-panel" width="100%">
-                <hbox alignSelf="END">
-                    <textField id="searchField" width="15em" placeholder="search project"/>
-                    <button id="searchButton" text="Search"/>
-                </hbox>
-                <button id="refreshButton" action="projectsDataGrid.refreshAction" alignSelf="CENTER"/>
+            <hbox width="100%">
+                <genericFilter id="genericFilter" dataLoader="projectsDl" themeNames="small"
+                               width="80%" alignSelf="START">
+                    <properties include=".*"/>
+                    <conditions> </conditions>
+                </genericFilter>
                 <simplePagination id="pagination" dataLoader="projectsDl" alignSelf="END"/>
             </hbox>
             <dataGrid id="projectsDataGrid"
                       width="100%"
-                      height="40em"
+                      height="30em"
                       dataContainer="projectsDc"
                       columnReorderingAllowed="true">
                 <actions>
                     <action id="createAction" type="list_create"/>
                     <action id="editAction" type="list_edit"/>
                     <action id="removeAction" type="list_remove"/>
-                    <action id="refreshAction" type="list_refresh"/>
                 </actions>
                 <columns resizable="true">
-                    <column property="projectName" header="Projectname"/>
+                    <column property="projectName" header="Name"/>
                 </columns>
             </dataGrid>
             <hbox id="buttonsPanel" classNames="buttons-panel">
