+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/vulnerability/vulnerability-detail-view.xml
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!--
+  ~ /*
+  ~ * Copyright (C) 2025 Bitsea GmbH
+  ~  *
+  ~ * Licensed under the Apache License, Version 2.0 (the "License");
+  ~ * you may not use this file except in compliance with the License.
+  ~ * You may obtain a copy of the License at
+  ~ *
+  ~ *     https://www.apache.org/licenses/LICENSE-2.0
+  ~ *
+  ~ * Unless required by applicable law or agreed to in writing, software
+  ~ * distributed under the License is distributed on an "AS IS" BASIS,
+  ~ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ * See the License for the specific language governing permissions and
+  ~ * limitations under the License.
+  ~ *
+  ~ * SPDX-License-Identifier: Apache-2.0
+  ~ * License-Filename: LICENSE
+  ~ */
+  ~
+  -->
+
+<view xmlns="http://jmix.io/schema/flowui/view"
+      title="msg://vulnerabilityDetailView.title">
+    <data>
+        <instance id="vulnerabilityDc"
+                  class="eu.occtet.bocfrontend.entity.Vulnerability">
+            <fetchPlan extends="_base"/>
+            <loader/>
+        </instance>
+        <collection id="inventoryItemsDc"
+                    class="eu.occtet.bocfrontend.entity.InventoryItem">
+            <fetchPlan extends="_base">
+                <property name="softwareComponent"/>
+                <property name="project" fetchPlan="_base">
+                    <property name="projectName"/>
+                </property>
+            </fetchPlan>
+            <loader id="inventoryItemsDl" readOnly="true">
+                <query>
+                    <![CDATA[select e from InventoryItem e]]>
+                </query>
+            </loader>
+        </collection>
+    </data>
+    <facets>
+        <dataLoadCoordinator auto="true"/>
+    </facets>
+    <actions>
+        <action id="saveAction" type="detail_saveClose"/>
+        <action id="closeAction" type="detail_close"/>
+    </actions>
+    <layout height="100%" width="100%">
+        <tabSheet id="mainTabSheet" width="100%" height="100%" themeNames="no-padding">
+            <tab id="detailTab" label="General information">
+                <vbox width="100%" height="100%">
+                    <formLayout id="form" dataContainer="vulnerabilityDc" >
+                        <textField id="vulnerabilityIdField" property="vulnerabilityId" label="Vulnerability ID"/>
+                        <hbox>
+                            <numberField id="weightedSeverityField" property="weightedSeverity" label="Weighted severity"/>
+                            <numberField id="exploitabilityField" property="exploitability" label="exploitability"/>
+                            <numberField id="riskScoreField" property="riskScore" label="Risk score"/>
+                        </hbox>
+                        <textArea id="summaryField" property="summary" label="Summary" height="6em"/>
+
+                        <textArea id="aliasesField" label="Aliases"/>
+                        <textField id="sourceUrlField" property="sourceUrl" label="Source URL"/>
+                        <textArea id="fixedPackagesField" label="Fixed Packages"/>
+                        <textArea id="referencesField" label="Reference URLs" height="25em"/>
+                    </formLayout>
+                    <hbox id="detailActions" spacing="true" alignSelf="CENTER" width="100%">
+                        <hbox alignSelf="START" width="40%">
+                            <span id="created" />
+                        </hbox>
+                        <hbox alignSelf="START" width="60%">
+                            <button id="commitAndCloseBtn" action="saveAction"/>
+                            <button id="closeBtn" action="closeAction"/>
+                        </hbox>
+                    </hbox>
+                </vbox>
+            </tab>
+            <tab id="findingTab" label="Findings">
+                <vbox>
+                    <hbox>
+                        <comboBox id="projectComboBox" label="Project"/>
+                        <comboBox id="softwareComboBox" label="SoftwareComponent"/>
+                    </hbox>
+                    <dataGrid id="inventoryItemsDataGrid"
+                              width="100%"
+                              height="40em"
+                              maxHeight="100%"
+                              dataContainer="inventoryItemsDc"
+                              columnReorderingAllowed="true">
+                        <columns resizable="true">
+                            <column property="project.projectName" header="Project"/>
+                            <column property="inventoryName" header="Inventory item"/>
+                            <column property="softwareComponent.name" header="Software component"/>
+                            <column property="softwareComponent.version" header="Version"/>
+                        </columns>
+                    </dataGrid>
+                </vbox>
+            </tab>
+        </tabSheet>
+    </layout>
+</view>
