+++ occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/vulnerability/vulnerability-list-view.xml
@@ -39,7 +39,6 @@
     <facets>
         <dataLoadCoordinator auto="true"/>
         <urlQueryParameters>
-            <genericFilter component="genericFilter"/>
             <pagination component="pagination"/>
         </urlQueryParameters>
     </facets>
@@ -48,17 +47,13 @@
         <action id="discardAction" type="lookup_discard"/>
     </actions>
     <layout>
-        <genericFilter id="genericFilter"
-                       dataLoader="vulnerabilitiesDl">
-            <properties include=".*"/>
-        </genericFilter>
-        <hbox id="buttonsPanel" classNames="buttons-panel">
-            <button id="editButton" action="vulnerabilitiesDataGrid.editAction"/>
-            <button id="excelExportButton" action="vulnerabilitiesDataGrid.excelExportAction"/>
-            <button id="jsonExportButton" action="vulnerabilitiesDataGrid.jsonExportAction"/>
-            <button id="refreshButton" action="vulnerabilitiesDataGrid.refreshAction"/>
-            <button id="removeButton" action="vulnerabilitiesDataGrid.removeAction"/>
-            <simplePagination id="pagination" dataLoader="vulnerabilitiesDl"/>
+        <hbox width="100%">
+            <genericFilter id="genericFilter" dataLoader="vulnerabilitiesDl" themeNames="small"
+                           width="80%" alignSelf="START">
+                <properties include=".*"/>
+                <conditions> </conditions>
+            </genericFilter>
+            <simplePagination id="pagination" dataLoader="vulnerabilitiesDl" alignSelf="END"/>
         </hbox>
         <dataGrid id="vulnerabilitiesDataGrid"
                   width="100%"
@@ -73,21 +68,29 @@
                 <action id="removeAction" type="list_remove"/>
             </actions>
             <columns resizable="true">
-                <column property="vulnerabilityId"/>
-                <column property="summary"/>
-                <column property="aliases"/>
-                <column property="weightedSeverity"/>
-                <column property="exploitability"/>
-                <column property="riskScore"/>
-                <column property="fixedPackages"/>
-                <column property="sourceUrl"/>
-                <column property="references"/>
-                <column property="createdAt"/>
+                <column property="vulnerabilityId" header="ID"/>
+                <column property="summary" header="Summary"/>
+                <column property="aliases" header="Aliases"/>
+                <column property="weightedSeverity" header="Serverity"/>
+                <column property="exploitability" header="Exploitability"/>
+                <column property="riskScore" header="Risk score"/>
+                <column property="fixedPackages" header="Fixed packages"/>
+                <column property="sourceUrl" header="Source url"/>
+                <column property="references" header="References"/>
+                <column property="createdAt" header="Created"/>
             </columns>
         </dataGrid>
         <hbox id="lookupActions" visible="false">
             <button id="selectButton" action="selectAction"/>
             <button id="discardButton" action="discardAction"/>
+        </hbox>
+        <hbox width="100%">
+            <button id="editButton" action="vulnerabilitiesDataGrid.editAction" alignSelf="START"/>
+            <button id="removeButton" action="vulnerabilitiesDataGrid.removeAction" alignSelf="START"/>
+            <hbox id="buttonsPanel" width="100%" justifyContent="CENTER">
+                <button id="excelExportButton" action="vulnerabilitiesDataGrid.excelExportAction" themeNames="primary"/>
+                <button id="jsonExportButton" action="vulnerabilitiesDataGrid.jsonExportAction" themeNames="primary"/>
+            </hbox>
         </hbox>
     </layout>
 </view>
