--- occtet-frontend/src/main/resources/eu/occtet/bocfrontend/view/vulnerabilitydetail/vulnerability-detail-view.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!--
-  ~ /*
-  ~ * Copyright (C) 2025 Bitsea GmbH
-  ~  *
-  ~ * Licensed under the Apache License, Version 2.0 (the "License");
-  ~ * you may not use this file except in compliance with the License.
-  ~ * You may obtain a copy of the License at
-  ~ *
-  ~ *     https://www.apache.org/licenses/LICENSE-2.0
-  ~ *
-  ~ * Unless required by applicable law or agreed to in writing, software
-  ~ * distributed under the License is distributed on an "AS IS" BASIS,
-  ~ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ * See the License for the specific language governing permissions and
-  ~ * limitations under the License.
-  ~ *
-  ~ * SPDX-License-Identifier: Apache-2.0
-  ~ * License-Filename: LICENSE
-  ~ */
-  ~
-  -->
-
-<view xmlns="http://jmix.io/schema/flowui/view"
-      title="msg://vulnerabilityDetailView.title">
-    <data>
-        <instance id="vulnerabilityDc"
-                  class="eu.occtet.bocfrontend.entity.Vulnerability">
-            <fetchPlan extends="_base"/>
-            <loader/>
-        </instance>
-    </data>
-    <facets>
-        <dataLoadCoordinator auto="true"/>
-    </facets>
-    <actions>
-        <action id="saveAction" type="detail_saveClose"/>
-        <action id="closeAction" type="detail_close"/>
-    </actions>
-    <layout alignItems="CENTER">
-        <formLayout id="form" dataContainer="vulnerabilityDc" >
-            <textField id="createdAtField" property="createdAt" label="Last updated" readOnly="true" width="AUTO"/>
-            <textField id="vulnerabilityIdField" property="vulnerabilityId" label="Vulnerability ID"/>
-            <textArea id="summaryField" property="summary" label="Summary" height="6em"/>
-
-            <textArea id="aliasesField" label="Aliases"/>
-            <hbox>
-                <numberField id="weightedSeverityField" property="weightedSeverity" label="Weighted severity"/>
-                <numberField id="exploitabilityField" property="exploitability" label="exploitability"/>
-                <numberField id="riskScoreField" property="riskScore" label="Risk score"/>
-            </hbox>
-            <textField id="sourceUrlField" property="sourceUrl" label="Source URL"/>
-            <textArea id="fixedPackagesField" label="Fixed Packages"/>
-            <textArea id="referencesField" label="Reference URLs"/>
-        </formLayout>
-        <hbox id="detailActions" spacing="true">
-            <button id="commitAndCloseBtn" action="saveAction"/>
-            <button id="closeBtn" action="closeAction"/>
-        </hbox>
-    </layout>
-</view>
